requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
pydantic==2.5.3
pydantic_core==2.14.6
pyTelegramBotAPI==4.15.4
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
telebot==0.0.5
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
